refactor: Penalty 도메인/애플리케이션 서비스 분리 및 책임 이동
