+++ backend/src/main/java/com/ice/studyroom/domain/penalty/application/PenaltyService.java
@@ -9,13 +9,12 @@ import java.time.LocalTime;
 import org.springframework.stereotype.Service;
 
 import com.ice.studyroom.domain.membership.domain.entity.Member;
-import com.ice.studyroom.domain.membership.domain.service.MemberDomainService;
 import com.ice.studyroom.domain.penalty.domain.entity.Penalty;
+import com.ice.studyroom.domain.penalty.domain.service.PenaltyDomainService;
 import com.ice.studyroom.domain.penalty.domain.type.PenaltyReasonType;
 import com.ice.studyroom.domain.penalty.domain.type.PenaltyStatus;
 import com.ice.studyroom.domain.penalty.infrastructure.persistence.PenaltyRepository;
 import com.ice.studyroom.domain.reservation.domain.entity.Reservation;
-import com.ice.studyroom.domain.reservation.domain.type.ReservationStatus;
 import com.ice.studyroom.domain.reservation.infrastructure.persistence.ReservationRepository;
 import com.ice.studyroom.global.exception.BusinessException;
 import com.ice.studyroom.global.type.StatusCode;
@@ -29,57 +28,28 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class PenaltyService {
 
+	private final PenaltyDomainService penaltyDomainService;
 	private final PenaltyRepository penaltyRepository;
 	private final ReservationRepository reservationRepository;
 
-	private final Clock clock;
-
 	@Transactional
 	public void assignPenalty(Member member, Long reservationId, PenaltyReasonType reason) {
+		Reservation reservation = reservationRepository.findById(reservationId).
+			orElseThrow(() -> new BusinessException(StatusCode.NOT_FOUND, "예약 정보를 찾을 수 없습니다."));
 
-		Penalty penalty = Penalty.builder()
-			.member(member)
-			.reservation(reservationRepository.findById(reservationId).get())
-			.reason(reason)
-			.penaltyEnd(calculatePenaltyEnd(reason.getDurationDays()))
-			.build();
-
+		Penalty penalty = penaltyDomainService.createPenalty(member, reservation, reason, null);
 		member.updatePenalty(true);
 		penaltyRepository.save(penalty);
 	}
 
 	public void adminAssignPenalty(Member member, LocalDateTime penaltyEndAt) {
-
-		Penalty penalty = Penalty.builder()
-			.member(member)
-			.reservation(null)
-			.reason(PenaltyReasonType.ADMIN)
-			.penaltyEnd(penaltyEndAt)
-			.build();
-
-		member.updatePenalty(true);
+		Penalty penalty = penaltyDomainService.createPenalty(member, null, PenaltyReasonType.ADMIN, penaltyEndAt);
 		penaltyRepository.save(penalty);
 	}
 
 	public void adminDeletePenalty(Member member) {
-		Penalty penalty = penaltyRepository.findByMemberIdAndStatus(member.getId(), PenaltyStatus.VALID).orElseThrow(
-			() -> new BusinessException(StatusCode.NOT_FOUND, "유효하지 않은 패널티입니다.")
-		);
+		Penalty penalty = penaltyDomainService.findPenaltyByMemberIdAndStatus(member);
 		penalty.expirePenalty();
 		member.updatePenalty(false);
-	}
-
-	private LocalDateTime calculatePenaltyEnd(int durationDays) {
-		LocalDate penaltyEndDate = LocalDate.now(clock);
-
-		int addedDays = 0;
-		while (addedDays < durationDays) {
-			penaltyEndDate = penaltyEndDate.plusDays(1);
-			if (!(penaltyEndDate.getDayOfWeek() == DayOfWeek.SATURDAY || penaltyEndDate.getDayOfWeek() == DayOfWeek.SUNDAY)) {
-				addedDays++;
-			}
-		}
-
-		return LocalDateTime.of(penaltyEndDate, LocalTime.of(23, 59, 59)); // 23:59:59
 	}
 }
